Fix CustomerInfo parsing and add support to non subscription parsing (#95)
## Motivation / Description
While I was testing the SDK I noticed we were not correctly creating the
CustomerInfo object. There was a problem mapping from the subscriptions
in the response to the final CustomerInfo. This fixes that and adds
support for parsing non subscriptions. We will not be supporting buying
them initially but I thought it would be good to still be able to access
the relevant data in the CustomerInfo object

## Changes introduced

## Linear ticket (if any)

## Additional comments