Validate email in SDK (#135)
## Motivation / Description
We were relying on the backend to validate the email. We can add some
initial validation in the SDK itself to make the response faster. Note
that we don't validate the email if provided by the developer, we will
still rely on the backend for that.

## Changes introduced

## Linear ticket (if any)

## Additional comments
<img width="1169" alt="image"
src="https://github.com/RevenueCat/purchases-js/assets/808417/88d418f0-b1b0-4342-b75a-24ef8c5b2fb5">