Track simple SDK_INITIALIZED event (#273)
## Description

Adds tracking for the SDK initialized event.

## Event Model

An event consists of:
- `id`: uuidv4 generated in the frontend for SDK events as soon as the
event happens. Note it is technically redundant given we also have the
`trace_id` and the `trace_index`.
- `type`: the type of event sent. Includes a prefix for Web Billing
`rcb_`. Follows same pattern as customer center events and paywall
events.
- `trace_id`: uuidv4 generated in the frontend as soon as the SDK is
initialized. All events of the same browser session will be sending this
same `trace_id` until the page is reloaded or exited.
- ~`trace_index`: an incremental 0-based index that describes the
partial order of the events in the same trace.~
- `timestamp`: an epoch in ms of when the event happened in the frontend
using the local clock. Follows same pattern as customer center events
and paywall events.

## Fake Timers

Vitest fake timers proofed a pretty powerful tool for testing the retry
logic of tracking events. It helps give determinism and reduce the
duration of the tests. No more arbitrary waits of X seconds.

## E2E

The e2e test currently fails given the backend does not accept the shape
of the event. It will be solved once the backend PR is merged.

## Retry Logic

The events are sent right away as they happen. If there is an unexpected
response from the browser or an error occurs with the request, the
events will resent in batch with an exponential back-off. Starts with 4s
and has a maximum interval of 60s. The back-off is reset upon successful
submission of the events.

⚠️ Currently the events are sent an infinite amount of times until the
request succeeds.
⚠️ Since events are fired right away if there have not been any errors
there may be multiple requests in-flight that send events to the backend
if none has yet failed.

---------

Co-authored-by: Pol Miro <polmiro@gmail.com>