[WEB-2053] Add an internal public function to track events from WPL (#292)
## Motivation / Description

WPL will be tracking some events that need to be in the same trace as
the rest of the `purchases-js` tracked events. There are two main
options:
1. Expose a `_trackEvent` public method marked as `@internal` for
`api-extractor`.
2. Expose the `_trace_id` public value marked as `@internal` for
`api-extractor`.

This pull requests implements the first option which seems like the best
one giving `@internal` tag is meant for such use cases and hides the
`trace_id` implementation detail.

> The @internal modifier is one of the four release tags. It indicates
that an API item is meant only for usage by other NPM packages from the
same maintainer. Third parties should never use "internal" APIs. To
emphasize this, an underscore prefix should be used in the name of a
declaration with an (explicit) @internal tag. API Extractor validates
this naming convention and will report
[ae-internal-missing-underscore](https://api-extractor.com/pages/messages/ae-internal-missing-underscore/)
if the underscore is missing.

See docs for
[here](https://api-extractor.com/pages/tsdoc/tag_internal/).

## Changes introduced

- Expose a public `_trackEvent` marked `@internal` on `Purchases` class.
- Run `npm audit fix`.