[WEB-2001] Add context to events (#298)
## Motivation / Description

Adds browser context to every event tracked. Based most of it from
Segment's track API.

Notice I decided to:
- Skipped `os` given it's parsed form `user_agent` so rather defer the
calculation to the backend for now.
- Skipped `ip` there is no standard way to fetch it from the browser so
might as well do it on the backend, plus it's going to be more
consistent to the way we gather those anyways in other places.