WEB-2065: Adding automatically utm params to metadata (#304)
## Motivation / Description
We want to track UTM parameters automatically in the WPL.
However this might be useful also for the sdk itself.


## Changes introduced
* Added a autoParseUTMParams function
* Added tests
* Using the autoParseUTMParams function before executing the purchase.

## Additional comments
This is an implicit behaviour, we could avoid doing this and passing
them only explicitly through the metadata (web-1998) but maybe is a good
dev experience if they get picked automatically without the need of
updating the code but just downloading the latest version of the sdk.