[WEB-2082] Append `trace_Id` and `checkout_session_id` to returned redeemUrl (#312)
## Motivation / Description

Given that we prefer to track the redemptions on the backend side, the
`redeem_url` will now include `trace_id` and `checkout_session_id` when
the url is a web one `http` or `https`. The future links will point to
Khepri and redirect the user to the custom scheme.

They could also be append them to mobile links with custom schemes but
given there is no expectation that they will make use of them I decided
to skip them. @tonidero let me know if you prefer me to add them for
those too.