Disable tax calculation E2E tests based on env var in CircleCI (#438)
## Motivation / Description
We are having troubles with Stripe tax calculation quotas and e2e tests.
This PR allows us to control if we run these tests through an env var in
CircleCI until we figure out if we can increase the quota.

## Changes introduced
Using an env var instead of the hardcoded value to enable e2e tests with
tax calculation