WEB-2370 - part 1: Refactored the Stripe Payment Elements component to isolate the Payment Elements part (#432)
## Motivation / Description
We need to add a custom Address Element and we can add the Email element
as well using the work from @polmiro in a previous PR.

This PR prepares the ground cleaning up the `stripe-elements` component
splitting the payment elements part in its dedicated component.

The next PRs will add the address and email components.

## Changes introduced
* Refactored the Stripe Payment Elements component to isolate the
Payment Elements part