[WEB-2450] Add PayButton with Wallet (#446)
## Motivation / Description

Adds some options to the PayButton to displaying the price for wallet
payments.

This is probably not using the exact spacings, and may need further
tweaks but it's a good place to get started. Since the new components
`InfoMessage`, the `PriceUpdateInfo` and the new properties of
`PayButton` are not used this is save to merge to main.

---------

Co-authored-by: Víctor Ferrer García <victorfergar@gmail.com>