Fix loading spinner positional glitch (#470)
## Motivation / Description

Fixes glitch with the loading spinner.

There are two different spinners being rendered:
1. on the initial load, waiting for the start of the checkout request to
go through.
2. on the payment entry page while stripe is loading, or the payment is
processing.

The second spinner was not well positioned. This is related to the
stripe form elements, which are kept on the screen with "visibility:
hidden" and not "display: none". If I recall well, it is to deal with
the animations that are provided by Stripe smoothly.

This PR uses "visibility: hidden" for the ".rc-checkout-form-container"
but uses "display:none" for the other parts that are not related to
Stripe Elements. Relies on "absolute" positioning for the loading so
that it will be centered on top of the "invsibile" Stripe Elements.