fix: save and apply original inline styles on close (#484)
## Motivation / Description
When closing the Purchases UI, several body styles are set to "auto",
which is not necessarily what they were set to prior to opening the
Purchase UI. The inline "auto" style overwrites styles defined via
classes, and also clobbers any inline styles that were previous set.

To be correct, it should save the inline style values in onMount and
then restore them after.

## Changes introduced
Original html and body styles are saved on mount and restored on destroy

## Linear ticket (if any)
Issue #415