[WEB-2510] Pass brandingAppearance to Buttons (#461)
## Motivation / Description

Linked with https://github.com/RevenueCat/rc-billing-checkout/pull/164.

We don't currently pass brandingAppearance to components - styles are
inherited through the theme set at the top level and the UI Library
Button Component just consumes these.

In the React Apps, since we render the new shared components/custom
elements in the shadow DOM, we need to pass the brandingAppearance to
them to create the styles via the Theme class. But if brandingAppearance
is empty, it will need to create default styles. However if we do that,
then when it's empty for the Svelte App then default styles will always
be applied.

Solution A: Add an "isSvelte/isReact" prop

I worry about this being a bit of a red flag, I don't want every shared
component to have a bunch of conditionals if possible

Solution B: Enforce always passing brandingAppearance, even in the
Svelte App

Sets a clearer differentiation of the library and enforces the
brandingAppearance API. Use context to avoid prop drilling it
everywhere.

## Changes introduced
- Upgrade to purchases-ui-js 1.0.2
- Pass the brandingAppearance prop to any shared components (i.e the
Button)

## Linear ticket (if any)

## Additional comments