WEB-2643: Do not show recurrent payment message for one-time purchases (#514)
## Motivation / Description
When purchasing a subscription we show a message explaining the user
that following renewals will be charged to the card they insert in the
payment step.

We should not show this message for one-time (consumable or not)
purchases, this PR removes that message when a consumable is being
purchased.

## Changes introduced
* Not showing the recurrent payment text when the purchase is not a
subscription.
* Added E2E tests to verify that.