Update fixture (#528)
## Motivation / Description

User analytics events are being pushed to the sandbox projects with the
wrong value in `tax_inclusive` because the fixture is missing it. I
thought events were being intercepted but seems this may not be the
case. Might be something related to CI but for now just updating the
fixture.