[WEB-2193] Add Intro Price Support (#536)
## Motivation / Description
This PR adds support for introductory pricing to the subscription
options in the purchases-js library. Users can now see introductory
prices if they are available, which helps them understand the costs
associated with their subscriptions better.

## Changes introduced
- Added a new property `introPrice` to the `SubscriptionOption`
interface and updated the API documentation to reflect this.
- Updated the sample web billing app to include introductory prices in
the UI, making it clearer to users.
- Adjusted typography and layout in the pricing summary to better
accommodate the new intro price information.
- Modified translation keys and strings to provide clear messaging
around introductory pricing.
- Ensured that API responses include introductory price information when
it exists.

## Linear ticket (if any)

## Additional comments
None

---------

Co-authored-by: James O'Donnell <james.odonnell@revenuecat.com>
Co-authored-by: Iga <iga@revenuecat.com>
Co-authored-by: Nicola Sacco <nicola@revenuecat.com>