Support test store (#560)
## Motivation / Description
This PR adds support for a new test store that can be used to test the
RC integration before actually having to setup the stores.
When a user tries to purchase a product obtained through a `test_` api
key, instead of going through the existing flow, we will display a
simple modal with some product info, and the developer can purchase it
using the test store. This can be useful to test the RC integration
without actually having to setup Web billing. Can be specially useful
for vibe coding tools.


https://github.com/user-attachments/assets/7f08b583-6adc-433f-92ff-c4ddc311541d



## Changes introduced

## Linear ticket (if any)

## Additional comments