Add internal API to postSimulatedStoreReceipt (#561)
## Motivation / Description
In this PR we add a new API to allow to post a purchase for a product in
the simulated store. This will be used by the React native SDK when
running in expo go mode, where it will need to post the receipt but
needs to handle the UI differently than how it's handled on browser
platforms.

Usage of this API requires to explicitely `// @ts-ignore` when calling
this method.

This API is not meant to be used by others, only meant for Internal
RevenueCat usage.

## Changes introduced

## Linear ticket (if any)

## Additional comments