Fix Expo Go compatibility (#585)
## Motivation / Description
We were running into issues in Expo Go in react-native-purchases because
it wasn't able to find the `crypto` library, and when importing the
`uuid` library it was saying that it couldn't find the crypto library.

This refactors the code to handle missing `crypto` definition + use our
custom code to generate UUIDs, instead of using the `UUID` library
directly, which fails because it internally uses `crypto` when importing
it.

Note that this means that in Expo Go, we will only have pseudo-random
UUID generation.

## Changes introduced

## Linear ticket (if any)

## Additional comments