Fix copy for single occurence of periodUnit (#611)
## Motivation / Description

Internal slack thread:
https://revenuecat.slack.com/archives/C03UKCQLW1Z/p1759759527117529

We currently say e.g "First 1 year for $14.99 paid once"

We should instead say "First year for $14.99 paid once", for the case
where the unit is `1`.

This is non-trivial to implement in our templating system, without
degrading translations.
We've agreed to carve out an exception for English language, given it
covers 95%+ of subscribers so makes sense to optimise. Broader planned
changes will eventually override these translations in the mid term.

## Changes introduced

## Linear ticket (if any)

## Additional comments