MON-1420: Populating information about the package offers to pass them to the Paywall  (#635)
## Motivation / Description
When an intro offer or a trial are defined the Paywall overrides certain
values, in particular in the Text component.
I just added the PackageInfo data structure to be passed in a Map per
package id to the Paywall ui component in a separate PR for
`purchases-ui-js`.

This PR converts the offering data we receive from the server into a
simple object as follows:
```
{ 'package_id': { hasTrial:boolean, hasIntroOffer:boolean}}
```

And passes it to the paywall.

## Changes introduced
* Bumped `purchases-ui-js` to 2.1.1
* Added the new `parseOfferingIntoPackageInfoPerPackage` function
* Used the function to populate the `infoPerPackage` parameter to be
passed to the Paywall
* Added some unit tests for the new function

---------

Co-authored-by: Marek Dabek <marek@revenuecat.com>