WEB-3392: Making offering an optional parameter for presentPaywall (#639)
## Motivation / Description
All the other sdks allow the offering to be a nullable/undefined
parameter when calling `presentPaywall`.
This PR implements the same logic for Web.
Whenever no offering is passed, we check if there's a paywall in
.current.
If not we raise an exception.

## Changes introduced
* Added the "pick .current when no offering passed"
* Updated the data PresentPaywallParams interface to make offering
optional
* Fixed the previously skipped E2E tests for paywalls
* Added a new E2E test to check that the paywall from the .current
offering of our E2E test project is picked.