Express purchase button support in paywalls (#704)
## Motivation
This PR aims to update the handling of the onReady callback in the
express purchase workflow, enhancing its functionality by including the
availability of wallets as part of the callback parameters. The changes
is needed to support the changes included in [this
PR](https://github.com/RevenueCat/purchases-ui-js/pull/196)
## Changes
* Updated the onReady callback in the ExpressPurchaseButton components
and related files to include walletsAvailable as a new parameter.
* Added logic to detect available payment methods (e.g., Apple Pay,
Google Pay).
* Adjusted main.ts and other relevant files to handle the new callback
parameter in the express purchase workflow.